# backend/api/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FeedTrack API",
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://feedtrack-backend.vercel.app",
      "description": "Production server"
    }
  ],
  "paths": {
    "/api/users": {
      "get": {
        "summary": "Get all users",
        "description": "Get a list of all users",
        "responses": {
          "200": {
            "description": "A list of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Error fetching users"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add a new user",
        "description": "Create a new admin",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Admin successfully created.",
            "items": {
              "type": "string",
              "example": "Admin successfully created."
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "enum": ["Missing required fields", "User already exists", "Invalid role"]
                        }
                      }
                    },
                    {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "example": "Missing required fields"
                      }
                    }
                  ]
                }
              }
            }
          },          
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Internal Server Error"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete all users",
        "description": "Delete all users from the system",
        "responses": {
          "204": {
            "description": "Users deleted successfully"
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{id}": {
      "delete": {
        "summary": "Delete user by id",
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64"
          },
          "description": "ID of the user to delete"
        }
      ],
      "responses": {
        "204": {
          "description": "User successfully deleted"
        },
        "500": {
          "description": "Internal server error",
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "message": {
                    "type": "string",
                    "example": "Error deleting user"
                  }
                }
              }
            }
          }
        }
      }
      }
    },
...
    "/api/feedbacks": {
      "get": {
        "summary": "Get all feedbacks",
        "description": "Get a list of all feedbacks",
        "responses": {
          "200": {
            "description": "A list of feedbacks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Feedback"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Error fetching feedbacks"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add a new feedback",
        "description": "Create a new feedback",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Feedback"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Feedback created successfully"
          },
          "400": {
            "description": "Bad request - Missing required fields",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Missing required fields"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Error creating feedback"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete all feedbacks",
        "description": "Delete all feedbacks from the system",
        "responses": {
          "204": {
            "description": "Feedbacks deleted successfully"
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Error deleting feedback:"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/branches": {
      "get": {
        "summary": "Get all branches",
        "description": "Get a list of all branches",
        "responses": {
          "200": {
            "description": "A list of branches",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Branch"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Error fetching branches"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add a new branch",
        "description": "Create a new branch",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Branch"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Branch created successfully"
          },
          "400": {
            "description": "Bad request - Missing required fields",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Missing required fields"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Error creating branch"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete all branches",
        "description": "Delete all branches from the system",
        "responses": {
          "204": {
            "description": "Branches deleted successfully"
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Error deleting branches"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/tellers": {
      "get": {
        "summary": "Get all tellers",
        "description": "Get a list of all tellers",
        "responses": {
          "200": {
            "description": "A list of tellers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Teller"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Error fetching tellers"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add a new teller",
        "description": "Create a new teller",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Teller"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Teller created successfully"
          },
          "400": {
            "description": "Bad request - Missing required fields",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Missing required fields"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Error creating teller"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete all tellers",
        "description": "Delete all tellers from the system",
        "responses": {
          "204": {
            "description": "Tellers deleted successfully"
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Error deleting tellers"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/questions": {
      "get": {
        "summary": "Get all questions",
        "description": "Get a list of all questions",
        "responses": {
          "200": {
            "description": "A list of questions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Question"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Error fetching questions"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add a new question",
        "description": "Create a new question",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Question"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Question created successfully"
          },
          "400": {
            "description": "Bad request - Missing required fields",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Missing required fields"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Error creating question"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete all questions",
        "description": "Delete all questions from the system",
        "responses": {
          "204": {
            "description": "Questions deleted successfully"
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Error deleting questions"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/campaigns": {
      "get": {
        "summary": "Get all campaigns",
        "description": "Get a list of all campaigns",
        "responses": {
          "200": {
            "description": "A list of campaigns",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Campaign"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Error fetching campaigns"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add a new campaign",
        "description": "Create a new campaign",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Campaign"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Campaign created successfully"
          },
          "400": {
            "description": "Bad request - Missing required fields",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Missing required fields"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Error creating campaign"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete all campaigns",
        "description": "Delete all campaigns from the system",
        "responses": {
          "204": {
            "description": "Campaigns deleted successfully"
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Error deleting campaigns"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/campaignQuestions": {
      "get": {
        "summary": "Get all campaign questions",
        "description": "Get a list of all campaign questions",
        "responses": {
          "200": {
            "description": "A list of campaign questions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CampaignQuestion"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Error fetching campaign questions"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add a new campaign question",
        "description": "Create a new campaign question",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CampaignQuestion"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Campaign question created successfully"
          },
          "400": {
            "description": "Bad request - Missing required fields",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Missing required fields"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Error creating campaign question"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete all campaign questions",
        "description": "Delete all campaign questions from the system",
        "responses": {
          "204": {
            "description": "Campaign questions deleted successfully"
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Error deleting campaign questions"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "int64",
            "example": 71
          },
          "name": {
            "type": "string",
            "example": "ime"
          },
          "lastname": {
            "type": "string",
            "example": "prezime"
          },
          "image": {
            "type": "string",
            "example": "/"
          },
          "password": {
            "type": "string",
            "example": "kj$sslk393vs&ls"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "ime@hotmail.com"
          },
          "mobilenumber": {
            "type": "string",
            "example": "062221111"
          },
          "role": {
            "type": "string",
            "example": "superAdmin"
          },
          "verified": {
            "type": "boolean",
            "example": true
          }
        }
      },      
      "Feedback": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "int64",
            "example": 71
          },
          "date": {
            "type": "string",
            "example": "2024-04-22T22:32:24.000Z"
          },
          "rating": {
            "type": "integer",
            "example": 3
          },
          "tellerPositionID": {
            "type": "string",
            "example": 15
          },
          "questionID": {
            "type": "string",
            "example": 28
          },
          "campaignID": {
            "type": "string",
            "example": 27
          }
        }
      },      
      "Branch": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "int64",
            "example": 19
          },
          "location": {
            "type": "string",
            "example": "Zenica"
          },
          "managerID": {
            "type": "string",
            "example": 1
          },
          "rating": {
            "type": "integer",
            "example": 4
          },
          "area": {
            "type": "string",
            "example": "Zeničko-dobojski kanton"
          },
          "campaignID": {
            "type": "string",
            "example": 2
          }
        }
      },
      "Teller": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": 1
          },
          "branchID": {
            "type": "string",
            "example": 2
          },
          "managerID": {
            "type": "string",
            "example": 6
          },
          "rating": {
            "type": "integer",
            "example": 4
          }
        }
      },
      "Question": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "What is your favorite color?"
          }
        }
      },
      "Campaign": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": 12
          },
          "name": {
            "type": "string",
            "example": "Graduation Gift Galore"
          },
          "startDate": {
            "type": "string",
            "format": "date-time",
            "example": null
          },
          "endDate": {
            "type": "string",
            "format": "date-time",
            "example": null
          },
          "questionsperpage": {
            "type": "string",
            "example": 2
          }
        }
      },
      "CampaignQuestion": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": 1
          },
          "campaignID": {
            "type": "string",
            "example": 1
          },
          "questionID": {
            "type": "string",
            "example": 1
          }
        }
      },
...
    }
  }
}